Split push logic into a separate extension to resolve dependency issues when using the module combined with google cloud functions Moved the dispatching logic to a separate generic class, so the vert.x extension only contains the vert.x specific logic
