+++ pubsub-push-vertx-http/deployment/pom.xml
@@ -2,14 +2,14 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <parent>
         <groupId>io.quarkiverse.googlecloudservices</groupId>
-        <artifactId>quarkus-google-cloud-pubsub-parent</artifactId>
+        <artifactId>quarkus-google-cloud-pubsub-push-vertx-http-parent</artifactId>
         <version>2.20.0-SNAPSHOT</version>
         <relativePath>../pom.xml</relativePath>
     </parent>
     <modelVersion>4.0.0</modelVersion>
 
-    <artifactId>quarkus-google-cloud-pubsub-deployment</artifactId>
-    <name>Quarkus - Google Cloud Services - PubSub - Deployment</name>
+    <artifactId>quarkus-google-cloud-pubsub-push-vertx-http-deployment</artifactId>
+    <name>Quarkus - Google Cloud Services - PubSub Push Vert.x HTTP Adapter - Deployment</name>
 
     <dependencies>
         <dependency>
@@ -17,24 +17,20 @@
             <artifactId>quarkus-core-deployment</artifactId>
         </dependency>
         <dependency>
-            <groupId>io.quarkus</groupId>
-            <artifactId>quarkus-grpc-common-deployment</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>io.quarkus</groupId>
-            <artifactId>quarkus-devservices-deployment</artifactId>
-        </dependency>
-        <dependency>
             <groupId>io.quarkiverse.googlecloudservices</groupId>
             <artifactId>quarkus-google-cloud-common-deployment</artifactId>
         </dependency>
         <dependency>
             <groupId>io.quarkiverse.googlecloudservices</groupId>
-            <artifactId>quarkus-google-cloud-pubsub</artifactId>
+            <artifactId>quarkus-google-cloud-pubsub-deployment</artifactId>
         </dependency>
         <dependency>
-            <groupId>org.testcontainers</groupId>
-            <artifactId>gcloud</artifactId>
+            <groupId>io.quarkiverse.googlecloudservices</groupId>
+            <artifactId>quarkus-google-cloud-pubsub-push-vertx-http</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>io.quarkus</groupId>
+            <artifactId>quarkus-vertx-http-deployment</artifactId>
         </dependency>
     </dependencies>
