+++ pubsub-push-vertx-http/runtime/pom.xml
@@ -2,60 +2,24 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <parent>
         <groupId>io.quarkiverse.googlecloudservices</groupId>
-        <artifactId>quarkus-google-cloud-pubsub-parent</artifactId>
+        <artifactId>quarkus-google-cloud-pubsub-push-vertx-http-parent</artifactId>
         <version>2.20.0-SNAPSHOT</version>
         <relativePath>../pom.xml</relativePath>
     </parent>
     <modelVersion>4.0.0</modelVersion>
 
-    <artifactId>quarkus-google-cloud-pubsub</artifactId>
-    <name>Quarkus - Google Cloud Services - PubSub - Runtime</name>
-    <description>Use Google Cloud PubSub messaging broker service</description>
+    <artifactId>quarkus-google-cloud-pubsub-push-vertx-http</artifactId>
+    <name>Quarkus - Google Cloud Services - PubSub Push Vert.x HTTP Adapter - Runtime</name>
+    <description>Use Google Cloud PubSub adapter logic for Vert.x HTTP based Push</description>
 
     <dependencies>
         <dependency>
             <groupId>io.quarkiverse.googlecloudservices</groupId>
-            <artifactId>quarkus-google-cloud-common</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>io.quarkiverse.googlecloudservices</groupId>
-            <artifactId>quarkus-google-cloud-common-grpc</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>com.google.cloud</groupId>
-            <artifactId>google-cloud-pubsub</artifactId>
-            <exclusions>
-                <exclusion>
-                    <groupId>commons-logging</groupId>
-                    <artifactId>commons-logging</artifactId>
-                </exclusion>
-                <exclusion>
-                    <groupId>javax.annotation</groupId>
-                    <artifactId>javax.annotation-api</artifactId>
-                </exclusion>
-                <exclusion>
-                    <groupId>org.checkerframework</groupId>
-                    <artifactId>checker-qual</artifactId>
-                </exclusion>
-                <!-- We exclude the grpc-netty-shaded library as grpc-netty is included via
-                quarkus-google-cloud-common-grpc and Quarkus includes the netty library -->
-                <exclusion>
-                    <groupId>io.grpc</groupId>
-                    <artifactId>grpc-netty-shaded</artifactId>
-                </exclusion>
-                <exclusion>
-                    <groupId>org.codehaus.mojo</groupId>
-                    <artifactId>animal-sniffer-annotations</artifactId>
-                </exclusion>
-            </exclusions>
+            <artifactId>quarkus-google-cloud-pubsub</artifactId>
         </dependency>
         <dependency>
             <groupId>io.quarkus</groupId>
-            <artifactId>quarkus-grpc-common</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>com.google.api-client</groupId>
-            <artifactId>google-api-client</artifactId>
+            <artifactId>quarkus-vertx-http</artifactId>
         </dependency>
     </dependencies>
