+++ pubsub-push-vertx-http/runtime/src/main/java/io/quarkiverse/googlecloudservices/pubsub/push/GooglePubSubAuthenticationHandler.java
