+++ pubsub-push-vertx-http/runtime/src/main/resources/META-INF/quarkus-extension.yaml
@@ -1,0 +1,16 @@
+---
+name: "Google Cloud Pubsub Push Vert.x HTTP Adapter"
+artifact: ${project.groupId}:${project.artifactId}:${project.version}
+metadata:
+  keywords:
+  - "pubsub"
+  - "google cloud"
+  - "gcloud"
+  - "gcp"
+  categories:
+  - "cloud"
+  - "messaging"
+  guide: "https://quarkiverse.github.io/quarkiverse-docs/quarkus-google-cloud-services/main/pubsub.html"
+  status: "preview"
+  config:
+  - "quarkus.google.cloud.pubsub.push"
