+++ pubsub/deployment/src/main/java/io/quarkiverse/googlecloudservices/pubsub/deployment/PubSubBuildSteps.java
@@ -5,12 +5,7 @@ import io.quarkiverse.googlecloudservices.pubsub.push.*;
 import io.quarkus.arc.deployment.AdditionalBeanBuildItem;
 import io.quarkus.deployment.annotations.BuildProducer;
 import io.quarkus.deployment.annotations.BuildStep;
-import io.quarkus.deployment.annotations.ExecutionTime;
-import io.quarkus.deployment.annotations.Record;
 import io.quarkus.deployment.builditem.FeatureBuildItem;
-import io.quarkus.vertx.http.deployment.FilterBuildItem;
-import io.quarkus.vertx.http.deployment.NonApplicationRootPathBuildItem;
-import io.quarkus.vertx.http.deployment.RouteBuildItem;
 
 public class PubSubBuildSteps {
     protected static final String FEATURE = "google-cloud-pubsub";
@@ -26,12 +21,7 @@ public class PubSubBuildSteps {
     }
 
     @BuildStep
-    @Record(ExecutionTime.RUNTIME_INIT)
     public void setupPush(BuildProducer<AdditionalBeanBuildItem> additionalBeans,
-            BuildProducer<RouteBuildItem> additionalRoutes,
-            BuildProducer<FilterBuildItem> additionalFilters,
-            NonApplicationRootPathBuildItem nonApplicationRootPathBuildItem,
-            PubSubPushRecorder recorder,
             PubSubPushBuildTimeConfig config) {
         if (!config.enabled()) {
             return;
@@ -43,19 +33,8 @@ public class PubSubBuildSteps {
                 .addBeanClasses(
                         PubSubPushManager.class,
                         GoogleTokenVerifier.class,
-                        PubSubPushEndpointHandler.class)
+                        MessageDispatcher.class)
                 .build());
-
-        config.endpointPath().ifPresent(endpoint -> {
-            additionalRoutes.produce(nonApplicationRootPathBuildItem
-                    .routeBuilder()
-                    .routeFunction(endpoint, recorder.routeFunction())
-                    .build());
-
-            additionalFilters.produce(new FilterBuildItem(
-                    recorder.authHandlerInstance(recorder.authenticationHandler(endpoint, config)),
-                    999));
-        });
 
     }
